3the maar goed

+++ Hack-the-future-2022-main/Htf2022.Deelbezems/Models/Message.cs
@@ -7,7 +7,7 @@
         public DateTime Timestamp { get; set; }
         public int StatusType { get; set; }
         public Positie Positie { get; set; }
-        public int BrandstofLevel { get; set; }
+        public double BrandstofLevel { get; set; }
 
         public string TaxiStatusId { get; set; }
         public string KastId { get; set; }
